Semantic Scholar Author Lookup

Obtain metadata relative to authors from Semantic Scholar including links to papers, citation profile and name variants. Please play nice with the API. If you need to access more data, there is also a data dump available. For more info see https://api.semanticscholar.org/corpus/.

s2_authors(s2_author_id = NULL, .progress = "none")

Arguments

s2_author_id
character, Semantic Scholar Author ID.

.progress
Shows the plyr-style progress bar. Options are "none", "text", "tk", "win", and "time". See create_progress_bar for details of each. By default, no progress bar is displayed.

Examples

if (FALSE) {
# one author
s2_authors("2204561")

# multiple authors at once with progress bar
s2_authors(c("2204561", "144128278", "49930593"), .progress = "text")
}