LSemantic Scholar Article Lookup

Obtain information about individual scholarly articles indexed in Semantic Scholar. Please play nice with the API. If you need to access more data, there is also a data dump available. For more info see https://api.semanticscholar.org/corpus/.

s2_papers(ids = NULL, .progress = "none")

Arguments

ids
character vector, get metadata by a single identifier or many identifiers. Besides its own id scheme, Semantic Scholar supports DOIs and arXiv IDs.

.progress
Shows the plyr-style progress bar. Options are "none", "text", "tk", "win", and "time". See create_progress_bar for details of each. By default, no progress bar is displayed.

Value

The result is a tibble with each row representing a publication. Here are the returned columns and descriptions according to the API docu:

abstract	Extracted abstract of the paper.

arxiv_id	arXiv ID

authors	List-column of detailed author infos

citation_velocity	A weighted average of the publication's citations for the last 3 years and fewer for publications published in the last year or two, which indicates how popular and lasting the publication is.

citations	lList-column of Semantic Scholar papers which cited this paper.

doi	Digital Object Identifier

influential_citation_count	Sum of the highly influential citations

paper_id	Semantic Scholar paper ID

reference	List-column of Semantic Scholar papers the paper cited

title	Article Title

topics	List-column of the fields of study this paper addresses.

url	Link article's landing page

venue	Extracted publication venue for this paper.

year	Publication year

Examples

if (FALSE) {
s2_papers("10.1186/1471-2164-11-245")

# call many papers with different id types at once
s2_papers(ids = c("10.1093/nar/gkr1047",
  "bbc25a6a340365832d4d27f683646c39f2661c88",
  "10.7717/peerj.2323",
  "arXiv:0711.0914")
  )
}